Update filter option to boogie-filter
